requirements.txt
flask==2.3.2
flask-cors==3.0.10
tensorflow-cpu==2.17.0
pillow==10.0.0
numpy==1.26.0
gunicorn==20.1.0
